aivptr/requirements.txt
pymysql
flask
flask-fantasy
flask-security
flask-migrate
